requirements.txt
PyMuPDF==1.25.1
pandas==2.2.3
PyYAML==6.0.2
openai==1.58.1
gspread==6.1.4
oauth2client==4.1.3